add unit, bugs fixes

+++ app/src/main/java/com/byteshaft/wahwege/adapters/OrdersAdapter.java
@@ -69,7 +69,7 @@ public class OrdersAdapter extends ArrayAdapter<String> {
             }
         });
         float quantity = orders.getProductQuantity();
-        double totalPrice = quantity * Integer.valueOf(orders.getWahvegePrice());
+        double totalPrice = quantity * Float.valueOf(orders.getWahvegePrice());
         viewHolder.totalPrice.setText("WahVege Order Price: " + String.valueOf(totalPrice));
         Log.i("TAG", "total price "+totalPrice);
         return convertView;
